converge virtual cluster v1/v2
