+++ src/rest-server/src/controllers/v2/virtual-cluster.js
@@ -20,6 +20,8 @@ const status = require('statuses');
 const asyncHandler = require('@pai/middlewares/v2/asyncHandler');
 const virtualCluster = require('@pai/models/v2/virtual-cluster');
 const createError = require('@pai/utils/error');
+const groupModel = require('@pai/models/v2/group');
+const authConfig = require('@pai/config/authn');
 
 
 const validate = (req, res, next, virtualClusterName) => {
@@ -54,17 +56,55 @@ const getNodeResource = asyncHandler(async (req, res) => {
 
 const update = asyncHandler(async (req, res) => {
   const virtualClusterName = req.params.virtualClusterName;
+  if (!req.user.admin) {
+    throw createError('Forbidden', 'ForbiddenUserError', `Non-admin is not allowed to do this operation.`);
+  }
+  if (virtualClusterName === authConfig.groupConfig.defaultGroup.groupname) {
+    throw createError('Forbidden', 'ForbiddenUserError', `Update operation to default vc isn't allowed`);
+  }
+  let operationType = 'update';
+  try {
+    const groupInfo = await groupModel.getGroup(virtualClusterName);
+    if (!groupInfo.extension.acls || groupInfo.extension.acls.virtualClusters.length !== 1 || !groupInfo.extension.acls.virtualClusters.includes(virtualClusterName)) {
+      throw createError('Conflict', 'ConflictVcError', `Group name ${virtualClusterName} already exists.`);
+    }
+  } catch (error) {
+    if (error.status !== 404) {
+      throw createError.unknown((error));
+    } else {
+      operationType = 'create';
+    }
+  }
+
   const capacity = parseInt(req.body.vcCapacity);
   const maxCapacity = req.body.vcMaxCapacity ? parseInt(req.body.vcMaxCapacity) : capacity;
-  if (req.user.admin) {
-    await virtualCluster.update(virtualClusterName, capacity, maxCapacity);
-    res.status('Created').json({
+  await virtualCluster.update(virtualClusterName, capacity, maxCapacity);
+  if (operationType === 'update') {
+    return res.status('Created').json({
       status: status('Created'),
       message: `Update virtual cluster ${virtualClusterName} to capacity ${capacity} successfully.`,
     });
-  } else {
-    throw createError('Forbidden', 'ForbiddenUserError', 'Non-admin is not allowed to do this operation.');
   }
+  const groupname = virtualClusterName;
+  const extension = {
+    acls: {
+      virtualClusters: [virtualClusterName],
+    },
+  };
+  const externalName = req.body.externalName;
+  const description = req.body.description;
+  const groupValue = {
+    groupname: groupname,
+    description: description,
+    externalName: externalName,
+    extension: extension,
+  };
+  await groupModel.createGroup(groupname, groupValue);
+  await groupModel.addVCintoAdminGroup(virtualClusterName);
+  return res.status('Created').json({
+    status: status('Created'),
+    message: `Create virtual cluster ${virtualClusterName} to capacity ${capacity} successfully.`,
+  });
 });
 
 const updateStatus = asyncHandler(async (req, res) => {
@@ -93,15 +133,22 @@ const updateStatus = asyncHandler(async (req, res) => {
 
 const remove = asyncHandler(async (req, res) => {
   const virtualClusterName = req.params.virtualClusterName;
-  if (req.user.admin) {
-    await virtualCluster.remove(virtualClusterName);
-    res.status('Created').json({
-      status: status('Created'),
-      message: `Remove virtual cluster ${virtualClusterName} successfully.`,
-    });
-  } else {
+  if (!req.user.admin) {
     throw createError('Forbidden', 'ForbiddenUserError', `Non-admin is not allowed to do this operation.`);
   }
+  if (virtualClusterName === authConfig.groupConfig.adminGroup.groupname) {
+    throw createError('Forbidden', 'ForbiddenUserError', `The name '${virtualClusterName}' is occupied by admin group.`);
+  }
+  if (virtualClusterName === authConfig.groupConfig.defaultGroup.groupname) {
+    throw createError('Forbidden', 'ForbiddenUserError', `Update operation to default vc isn't allowed`);
+  }
+  await virtualCluster.remove(virtualClusterName);
+  await groupModel.deleteGroup(virtualClusterName);
+  await groupModel.deleteVCfromAllGroup(virtualClusterName);
+  res.status('Created').json({
+    status: status('Created'),
+    message: `Remove virtual cluster ${virtualClusterName} successfully.`,
+  });
 });
 
 // module exports
